refactor: split auth oauth protocol helpers
